requirements_docker.txt
numpy
pandas
opencv-python
networkx
h5py
tqdm
vocab
revtok
Pillow
tensorboardX
ai2thor~=4.2.0
pyyaml
matplotlib
transformers
scikit-learn
torchmetrics
torchtext==0.12.0
graphviz
pydot
sentencepiece
nltk
sentence-transformers
open_clip_torch
git+https://github.com/openai/CLIP.git
